Refactor service.rs into smaller modules (#405)
* Refactor service.rs into smaller modules

* make substrate related crates feature flagged

* Update the imports

* Split the services by the system

* Update Cargo.lock
