+++ crates/tx-queue/Cargo.toml
@@ -22,8 +22,8 @@ futures = { workspace = true }
 backoff = { workspace = true }
 tokio = { workspace = true }
 webb = { workspace = true }
-sp-core = { workspace = true }
-sp-runtime = { workspace = true }
+sp-core = { workspace = true, optional = true }
+sp-runtime = { workspace = true, optional = true }
 # Used by ethers (but we need it to be vendored with the lib).
 native-tls = { workspace = true, optional = true }
 ethereum-types = { workspace = true }
@@ -34,9 +34,5 @@ rand = { version = "0.8", default-features = false, features = ["getrandom"] }
 [features]
 default = ["std", "evm", "substrate"]
 std = []
-evm = [
-    "webb-relayer-context/evm",
-]
-substrate = [
-    "webb-relayer-context/substrate",
-]
+evm = ["webb-relayer-context/evm"]
+substrate = ["webb-relayer-context/substrate", "sp-core", "sp-runtime"]
