+++ services/block-poller/src/lib.rs
@@ -3,7 +3,7 @@ use std::sync::Arc;
 use crate::block_poller::{BlockPoller, BlockPollingHandler};
 use ethereum_types::U256;
 use webb::evm::ethers::types::{Block, TxHash};
-use webb_relayer::service::{Client, Store};
+use webb_relayer::service::{evm::Client, Store};
 use webb_relayer_config::block_poller::BlockPollerConfig;
 use webb_relayer_context::RelayerContext;
 use webb_relayer_store::SledStore;
